Add support for setting alluxio master hostname for Mesos
