+++ integration/bin/alluxio-mesos.sh
@@ -58,7 +58,7 @@ mkdir -p "${ALLUXIO_LOGS_DIR}"
   -Dalluxio.home="${ALLUXIO_HOME}" \
   -Dalluxio.logger.type="Console" \
   -Dalluxio.logs.dir="${ALLUXIO_LOGS_DIR}" \
-  alluxio.mesos.AlluxioFramework "${MESOS_MASTER_ADDRESS}" > "${ALLUXIO_LOGS_DIR}"/framework.out 2>&1 &
+  alluxio.mesos.AlluxioFramework --mesos "${MESOS_MASTER_ADDRESS}" > "${ALLUXIO_LOGS_DIR}"/framework.out 2>&1 &
 
 if [[ "${wait}" ]]; then
   wait
