+++ integration/mesos/pom.xml
@@ -58,6 +58,10 @@
       <artifactId>alluxio-underfs-hdfs</artifactId>
       <version>${project.version}</version>
     </dependency>
+    <dependency>
+      <groupId>com.beust</groupId>
+      <artifactId>jcommander</artifactId>
+    </dependency>
 
     <!-- Other projects' test-jars -->
     <dependency>
